Files restructured and broken into multiple components for better readability
