+++ src/scripts.js
@@ -1,9 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Login from "./Components/Login";
+import Login from "./Components/Auth/Login";
 import Homepage from "./Components/Homepage/Homepage";
-import SignUp from "./Components/SignUp";
+import SignUp from "./Components/Auth/SignUp";
 import SeeTransactions from "./Components/Homepage/SeeTransactions";
 
 const Scripts = () => {
